backend/pyproject.toml
[project]
name = "enterprise-rag-backend"
version = "1.0.0"
description = "GDPR-compliant Enterprise RAG Assistant with LangGraph agentic workflows"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    # LangChain + LangGraph
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.3.0",
    "langgraph>=0.2.0",
    # Vector store
    "weaviate-client>=4.9.0",
    # Document processing
    "pypdf>=5.0.0",
    "markdown>=3.7",
    # Embeddings & tokens
    "tiktoken>=0.8.0",
    # Language detection
    "langdetect>=1.0.9",
    # Config
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    # CORS
    "httpx>=0.28.0",
    "langchain-text-splitters>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
